fix(query): Fix _docID-only query returning all fields and resolve clippy warnings
The build_mapping() function incorrectly treated queries requesting only
_docID as "no fields specified" and expanded all collection fields into
the response. This caused TestMutationUpdateOneToMany_AliasRelationName
AndInternalIDBothProduceSameDocID to fail with 5 fields returned instead
of 1.

Fix: add render_keys.is_empty() check so the "SELECT *" fallback only
triggers when truly no fields are requested.

Also resolves all clippy warnings across the workspace:
- Bump MSRV from 1.75 to 1.82 (code already used LazyLock requiring 1.80+)
- Remove 19 unnecessary u64 casts
- Use const thread_local initializers
- Derive Default instead of manual impls
- Replace manual prefix stripping with strip_prefix()
- Simplify boolean expressions, collapse nested if-let/match
- Add # Safety docs to FFI unsafe functions
- Various other lint fixes across ~50 files

FFI mutation tests: 200 pass, 1 fail (known SSI issue #244), 21 skip

+++ crates/query/src/plan/similarity.rs
@@ -160,7 +160,7 @@ impl PlanNode for SimilarityNode {
 
         obj.insert(
             "iterations".to_string(),
-            serde_json::json!(self.exec_info.iterations as u64),
+            serde_json::json!(self.exec_info.iterations),
         );
 
         let child_explain = self.source.explain_execute();
